refactor: implement context
