+++ packages/app-store/templates/audit-log-implementation/components/NavigationPanel.tsx
@@ -1,8 +1,10 @@
 import { VerticalTabs } from "@calcom/ui";
 
+import { useAppCredential } from "../context/CredentialContext";
 import { getAppSettingsOptions } from "../lib";
 
-export const NavigationPanel = (props: { credentialId: number }) => {
-  const appSettingsOptions = getAppSettingsOptions(props.credentialId);
+export const NavigationPanel = () => {
+  const { credentialId } = useAppCredential();
+  const appSettingsOptions = getAppSettingsOptions(credentialId);
   return <VerticalTabs tabs={appSettingsOptions} linkShallow itemClassname="w-[100%] h-[100%]" />;
 };
